finished employee implementation and started schedule logic

+++ src/app/components/delete-modal/delete-modal.component.ts
@@ -7,6 +7,8 @@ import {
 import { MatButtonModule } from '@angular/material/button';
 import { MAT_DIALOG_DATA, MatDialogModule } from '@angular/material/dialog';
 import { CustomerService } from '../../services/customer.service';
+import { Router } from '@angular/router';
+import { EmployeeService } from '../../services/employee.service';
 
 @Component({
   selector: 'app-delete-modal',
@@ -17,12 +19,27 @@ import { CustomerService } from '../../services/customer.service';
 })
 export class DeleteModalComponent {
   customerService: CustomerService = inject(CustomerService);
+  employeeService: EmployeeService = inject(EmployeeService);
 
   constructor(@Inject(MAT_DIALOG_DATA) public data: { id: string }) {}
 
+  router: Router = new Router();
+
+  ngOnInit() {
+    console.log(this.router.url);
+  }
+
   deleteItem() {
-    this.customerService.deleteCustomer(this.data.id).subscribe({
-      error: (err) => console.log(err),
-    });
+    if (this.router.url == '/customers') {
+      this.customerService.deleteCustomer(this.data.id).subscribe({
+        error: (err) => console.log(err),
+      });
+    }
+
+    if (this.router.url == '/employees') {
+      this.employeeService.deleteEmployee(this.data.id).subscribe({
+        error: (err) => console.log(err),
+      });
+    }
   }
 }
